requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
fastapi==0.115.0
Flask==3.0.3
flatbuffers==24.3.25
h11==0.14.0
humanfriendly==10.0
idna==3.10
imageio==2.35.1
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
lazy_loader==0.4
llvmlite==0.43.0
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numba==0.60.0
numpy==2.0.2
onnxruntime==1.19.2
opencv-python-headless==4.10.0.84
packaging==24.1
pillow==10.4.0
platformdirs==4.3.6
pooch==1.8.2
protobuf==5.28.2
pydantic==2.9.2
pydantic_core==2.23.4
PyMatting==1.1.12
pyreadline3==3.5.4
python-multipart==0.0.10
referencing==0.35.1
rembg==2.0.59
requests==2.32.3
rpds-py==0.20.0
scikit-image==0.24.0
scipy==1.14.1
setuptools==75.1.0
sniffio==1.3.1
starlette==0.38.6
sympy==1.13.3
tifffile==2024.9.20
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.30.6
Werkzeug==3.0.4
wheel==0.44.0
